Update the bukkit.yml to add the default world + generator.
- Block movement outside world height limits.
- Refactor /terralayers init to be more readable.
- Added Terraplusminus as plugin dependency.
- Updated world switching to use actual y level and accounting for the world offset.

+++ terralayers-api/src/main/java/org/btuk/terralayers/api/LayerManager.java
@@ -19,7 +19,7 @@ public interface LayerManager {
      *
      * @param globalY the y-level to get the layer for
      */
-    Optional<LayeredWorld> getLayerForGlobalY(int globalY);
+    Optional<LayeredWorld> getLayerForGlobalY(double globalY);
 
     /**
      * Finds the layer that matches the given Bukkit world.
